requirements.txt
duckdb
pandas
numpy
requests
beautifulsoup4
lxml
python-dotenv
faiss-cpu
sentence-transformers
tqdm
dart-fss 
opendartreader
streamlit